+++ CHANGES.md
@@ -18,6 +18,7 @@ Release Notes.
 * Force the injected high-priority classes in order to avoid NoClassDefFoundError.
 * Plugin to support xxl-job 2.3.x.
 * Add plugin to support Micronaut(HTTP Client/Server) 3.2.x-3.6.x
+* Add plugin to support NATS Java client 2.14.x-2.15.x
 
 #### Documentation